# openapi.yaml
openapi: 3.1.0
info:
  title: BASidekick BAS Data API
  version: 1.0.0
  description: Thin API wrapper for BAS Babel and BAS Atlas datasets.
servers:
  - url: /api
paths:
  /meta:
    get:
      summary: Dataset metadata
      responses:
        '200':
          description: Metadata response
  /points/{id}:
    get:
      summary: Get point by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: includeGraph
          schema:
            type: boolean
        - in: query
          name: depth
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - in: query
          name: rel
          schema:
            type: string
      responses:
        '200':
          description: Point details
        '404':
          description: Not found
  /equipment/{id}:
    get:
      summary: Get equipment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Equipment details
        '404':
          description: Not found
  /search:
    get:
      summary: Search Babel and Atlas resources
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
        '400':
          description: Missing query
  /normalize:
    post:
      summary: Normalize a point name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Normalization result
        '400':
          description: Invalid payload
  /validate:
    post:
      summary: Validate a point list against an equipment template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [equipmentTypeId, points]
              properties:
                equipmentTypeId:
                  type: string
                points:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Validation output
        '400':
          description: Invalid payload
  /templates/{equipmentTypeId}:
    get:
      summary: Get templates by equipment type
      parameters:
        - in: path
          name: equipmentTypeId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Matching templates
        '404':
          description: Not found